feat: add pure nix setup
Co-authored-by: Ulrik Strid <ulrik.strid@outlook.com>

remove ligo

Ligo doesn't bulid on any platform except x86-64_linux.
We'll disable it for now. Nix users on Linux can always
grab it with 'nix-shell -p ligo'.

remove docker-compose

We specify version 3.8 in our docker-compose file.
We should not be using the python docker-compose
pacakge, as the last version was 1.29